refactoring: moved idfa stuff to attribution fetcher. fixed tests
