+++ Purchases/Attribution/RCAttributionFetcher.m
@@ -49,37 +49,12 @@ static NSMutableArray<RCAttributionData *> *_Nullable postponedAttributionData;
     return self;
 }
 
-- (NSString *)rot13:(NSString *)string {
-    NSMutableString *rotatedString = [NSMutableString string];
-    for (NSUInteger charIdx = 0; charIdx < string.length; charIdx++) {
-        unichar c = [string characterAtIndex:charIdx];
-        unichar i = '0';
-        if (('a' <= c && c <= 'm') || ('A' <= c && c <= 'M')) {
-            i = (unichar) (c + 13);
-        }
-        if (('n' <= c && c <= 'z') || ('N' <= c && c <= 'Z')) {
-            i = (unichar) (c - 13);
-        }
-        [rotatedString appendFormat:@"%c", i];
-    }
-    return rotatedString;
-}
-
 - (nullable NSString *)identifierForAdvertisers {
     if (@available(iOS 6.0, macOS 10.14, *)) {
-        // We need to do this mangling to avoid Kid apps being rejected for getting idfa.
-        // It looks like during the app review process Apple does some string matching looking for
-        // functions in the AdSupport.framework. We apply rot13 on these functions and classes names
-        // so that Apple can't find them during the review, but we can still access them on runtime.
-        NSString *mangledClassName = @"NFVqragvsvreZnantre";
-        NSString *mangledIdentifierPropertyName = @"nqiregvfvatVqragvsvre";
-
-        NSString *className = [self rot13:mangledClassName];
-        id <FakeASIdentifierManager> asIdentifierManagerClass = (id <FakeASIdentifierManager>) NSClassFromString(className);
+        Class <FakeASIdentifierManager> _Nullable asIdentifierManagerClass = [self.attributionFactory asIdentifierClass];
         if (asIdentifierManagerClass) {
-            NSString *identifierPropertyName = [self rot13:mangledIdentifierPropertyName];
             id sharedManager = [asIdentifierManagerClass sharedManager];
-            NSUUID *identifierValue = [sharedManager valueForKey:identifierPropertyName];
+            NSUUID *identifierValue = [sharedManager valueForKey:[self.attributionFactory asIdentifierPropertyName]];
             return identifierValue.UUIDString;
         } else {
             RCWarnLog(@"%@", RCStrings.configure.adsupport_not_imported);
