+++ PurchasesTests/Mocks/MockAttributionTypeFactory.swift
@@ -32,7 +32,7 @@ class MockAdClient: NSObject, FakeAdClient {
 
 @available(iOS 14, macOS 11, tvOS 14, *)
 class MockATTrackingManager: NSObject, FakeATTrackingManager {
-    static var mockAuthorizationStatus: ATTrackingManager.AuthorizationStatus = .denied
+    static var mockAuthorizationStatus: ATTrackingManager.AuthorizationStatus = .authorized
 
     static func trackingAuthorizationStatus() -> Int {
         return Int(mockAuthorizationStatus.rawValue)
@@ -49,7 +49,7 @@ class MockAttributionTypeFactory: AttributionTypeFactory {
     static var shouldReturnTrackingManagerClass = true
 
     override func trackingManagerClass() -> FakeATTrackingManager.Type? {
-        if #available(iOS 14, *) {
+        if #available(iOS 14, macOS 11, tvOS 14, *) {
             return Self.shouldReturnTrackingManagerClass ? MockATTrackingManager.self : nil
         } else {
             return nil
