+++ PurchasesTests/PurchasesTests-Bridging-Header.h
@@ -17,7 +17,7 @@
 #include <Purchases/RCPackage.h>
 #include <Purchases/RCIntroEligibility+Protected.h>
 #include <Purchases/RCAttributionFetcher.h>
-#include <Purchases/RCAttributionFetcher+Protected.h>
+#include <Purchases/RCAttributionTypeFactory+Protected.h>
 #include <Purchases/RCAttributionData.h>
 #include <Purchases/RCPromotionalOffer+Protected.h>
 #include <Purchases/RCOfferingsFactory.h>
